DrataUIAutomation/requirements.txt
allure-behave==2.8.40
allure-python-commons==2.8.40
behave==1.2.6
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
pytzdata==2020.1
requests==2.25.0
selenium==3.141.0
uritemplate==3.0.1
urllib3==1.26.1
webdriver-manager==3.2.2
